requirements.txt
asgiref==3.3.4
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
